Delete .env
