+++ .nyc_output/processinfo/70a4f2f2-8948-4526-97e2-e353997bc0c7.json
@@ -1,0 +1,1 @@
+{"uuid":"70a4f2f2-8948-4526-97e2-e353997bc0c7","parent":null,"pid":23297,"argv":["/usr/bin/node","/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/node_modules/.bin/mocha","--exit"],"execArgv":[],"cwd":"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI","time":1561723559464,"ppid":23275,"root":"c38abc4c-5d6a-4c40-8f63-ab1c8e16438b","coverageFilename":"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/.nyc_output/70a4f2f2-8948-4526-97e2-e353997bc0c7.json","files":[]}
