--- .nyc_output/processinfo/93441098-2c3e-4737-847d-9abcf238f81a.json
@@ -1,1 +1,0 @@
-{"uuid":"93441098-2c3e-4737-847d-9abcf238f81a","parent":null,"pid":21114,"argv":["/usr/bin/node","/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/node_modules/.bin/mocha","--exit"],"execArgv":[],"cwd":"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI","time":1561633624987,"ppid":21106,"root":"0f644fb2-a656-4c0d-a466-9d0d180224b3","coverageFilename":"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/.nyc_output/93441098-2c3e-4737-847d-9abcf238f81a.json","files":[]}
