+++ .nyc_output/processinfo/a3c0a058-7932-4747-b246-5d40f78ac72a.json
@@ -1,0 +1,1 @@
+{"uuid":"a3c0a058-7932-4747-b246-5d40f78ac72a","parent":"70a4f2f2-8948-4526-97e2-e353997bc0c7","pid":23320,"argv":["/usr/bin/node","/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/node_modules/mocha/bin/_mocha","--exit","--no-config","--no-package","--no-opts","--diff","--extension","js","--reporter","spec","--slow","75","--timeout","2000","--ui","bdd"],"execArgv":[],"cwd":"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI","time":1561723562544,"ppid":23297,"root":"c38abc4c-5d6a-4c40-8f63-ab1c8e16438b","coverageFilename":"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/.nyc_output/a3c0a058-7932-4747-b246-5d40f78ac72a.json","files":["/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/models/index.js","/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/app.js","/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v1/models/index.js","/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/controllers/index.js","/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/controllers/properties/index.js","/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/controllers/users/index.js","/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/controllers/users/config.js","/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/controllers/users/middleware.js"]}
