--- .nyc_output/processinfo/c0345dba-51b6-458c-8918-2c90c546a593.json
@@ -1,1 +1,0 @@
-{"uuid":"c0345dba-51b6-458c-8918-2c90c546a593","parent":"93441098-2c3e-4737-847d-9abcf238f81a","pid":21120,"argv":["/usr/bin/node","/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/node_modules/mocha/bin/_mocha","--exit","--no-config","--no-package","--no-opts","--diff","--extension","js","--reporter","spec","--slow","75","--timeout","2000","--ui","bdd"],"execArgv":[],"cwd":"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI","time":1561633628221,"ppid":21114,"root":"0f644fb2-a656-4c0d-a466-9d0d180224b3","coverageFilename":"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/.nyc_output/c0345dba-51b6-458c-8918-2c90c546a593.json","files":["/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v1/models/index.js","/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/app.js","/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v1/controllers/index.js","/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v1/controllers/properties/index.js","/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v1/controllers/users/index.js","/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v1/controllers/users/config.js","/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v1/controllers/users/middleware.js","/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/controllers/index.js","/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/models/index.js","/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/controllers/properties/index.js","/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/controllers/users/index.js","/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/controllers/users/config.js","/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/controllers/users/middleware.js"]}
