+++ .nyc_output/processinfo/index.json
@@ -1,1 +1,1 @@
-{"processes":{"93441098-2c3e-4737-847d-9abcf238f81a":{"parent":null,"children":["c0345dba-51b6-458c-8918-2c90c546a593"]},"c0345dba-51b6-458c-8918-2c90c546a593":{"parent":"93441098-2c3e-4737-847d-9abcf238f81a","children":[]}},"files":{"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v1/models/index.js":["c0345dba-51b6-458c-8918-2c90c546a593"],"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/app.js":["c0345dba-51b6-458c-8918-2c90c546a593"],"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v1/controllers/index.js":["c0345dba-51b6-458c-8918-2c90c546a593"],"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v1/controllers/properties/index.js":["c0345dba-51b6-458c-8918-2c90c546a593"],"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v1/controllers/users/index.js":["c0345dba-51b6-458c-8918-2c90c546a593"],"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v1/controllers/users/config.js":["c0345dba-51b6-458c-8918-2c90c546a593"],"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v1/controllers/users/middleware.js":["c0345dba-51b6-458c-8918-2c90c546a593"],"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/controllers/index.js":["c0345dba-51b6-458c-8918-2c90c546a593"],"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/models/index.js":["c0345dba-51b6-458c-8918-2c90c546a593"],"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/controllers/properties/index.js":["c0345dba-51b6-458c-8918-2c90c546a593"],"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/controllers/users/index.js":["c0345dba-51b6-458c-8918-2c90c546a593"],"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/controllers/users/config.js":["c0345dba-51b6-458c-8918-2c90c546a593"],"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/controllers/users/middleware.js":["c0345dba-51b6-458c-8918-2c90c546a593"]},"externalIds":{}}
+{"processes":{"70a4f2f2-8948-4526-97e2-e353997bc0c7":{"parent":null,"children":["a3c0a058-7932-4747-b246-5d40f78ac72a"]},"a3c0a058-7932-4747-b246-5d40f78ac72a":{"parent":"70a4f2f2-8948-4526-97e2-e353997bc0c7","children":[]}},"files":{"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/models/index.js":["a3c0a058-7932-4747-b246-5d40f78ac72a"],"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/app.js":["a3c0a058-7932-4747-b246-5d40f78ac72a"],"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v1/models/index.js":["a3c0a058-7932-4747-b246-5d40f78ac72a"],"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/controllers/index.js":["a3c0a058-7932-4747-b246-5d40f78ac72a"],"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/controllers/properties/index.js":["a3c0a058-7932-4747-b246-5d40f78ac72a"],"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/controllers/users/index.js":["a3c0a058-7932-4747-b246-5d40f78ac72a"],"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/controllers/users/config.js":["a3c0a058-7932-4747-b246-5d40f78ac72a"],"/home/hezron/ANDELA-KIGALI/PropertyProLiteAPI/api/v2/controllers/users/middleware.js":["a3c0a058-7932-4747-b246-5d40f78ac72a"]},"externalIds":{}}
