+++ .travis.yml
@@ -9,6 +9,6 @@ script:
   - npm test
 before_script:
   - npm install jsonwebtoken
-  - psql -c 'CREATE DATABASE test;' -U postgres
+  - psql -c 'CREATE DATABASE ppltest;' -U postgres
 after_success:
   - npm install coveralls && npm run coverage
