+++ dbb.js
@@ -1,0 +1,67 @@
+const { Pool } = require('pg');
+const dotenv = require('dotenv');
+
+dotenv.config();
+
+const pool = new Pool({
+  connectionString: process.env.DATABASE_URL
+});
+
+pool.on('connect', () => {
+  console.log('connected to the db');
+});
+
+/**
+ * Create Tables
+ */
+const createTables = () => {
+  const queryText =
+    `CREATE TABLE IF NOT EXISTS
+      users(
+        firstName VARCHAR NOT NULL,
+        secondName VARCHAR NOT NULL,
+        username VARCHAR NOT NULL,
+        email VARCHAR NOT NULL,
+        phoneNumber VARCHAR NOT NULL,
+        password VARCHAR NOT NULL
+      )`;
+
+
+  pool.query(queryText)
+    .then((res) => {
+      console.log(res);
+      pool.end();
+    })
+    .catch((err) => {
+      console.log(err);
+      pool.end();
+    });
+}
+
+/**
+ * Drop Tables
+ */
+const dropTables = () => {
+  const queryText = 'DROP TABLE IF EXISTS reflections';
+  pool.query(queryText)
+    .then((res) => {
+      console.log(res);
+      pool.end();
+    })
+    .catch((err) => {
+      console.log(err);
+      pool.end();
+    });
+}
+
+pool.on('remove', () => {
+  console.log('client removed');
+  process.exit(0);
+});
+
+module.exports = {
+  createTables,
+  dropTables
+};
+
+require('make-runnable');
