+++ dbtest/properties.json
@@ -1,1 +1,27 @@
-[]
+[]  {
+    "category": "Sroom",
+    "name": "Houscbe",
+    "reason": "rencbct",
+    "price": "7000",
+    "state": "tancbania",
+    "city": "arha",
+    "address": "76768",
+    "map": "90800,800",
+    "description": "Veryl house",
+    "url": "https://res.cloudinary.com/hezzie/image/upload/v1561036548/PropertyProLiteAPI/2019-06-20T13:15:46.472Z.png",
+    "id": 2895
+  },
+  {
+    "category": "Sroom",
+    "name": "Houscbe",
+    "reason": "rencbct",
+    "price": "7000",
+    "state": "tancbania",
+    "city": "arha",
+    "address": "76768",
+    "map": "90800,800",
+    "description": "Veryl house",
+    "url": "https://res.cloudinary.com/hezzie/image/upload/v1561036548/PropertyProLiteAPI/2019-06-20T13:15:46.472Z.png",
+    "id": 229
+  }
+]
