+++ dbtest/users.json
@@ -1,12 +1,1 @@
-[
-  {
-    "firstName": "hezron",
-    "secondName": "kimutai",
-    "userName": "hezzie",
-    "email": "hezgmail.com",
-    "phoneNumber": "0937892356",
-    "password": "req",
-    "profilePic": "",
-    "id": 8672
-  }
-]
+[]
