+++ test/v1/users/index.js
@@ -1,6 +1,7 @@
 const records = require('../../../api/v1/models');
 const chai = require('chai');
 const server = require('../../../app');
+
 const user = {
     "firstName": "hezron",
     "secondName": "kimutai",
@@ -14,21 +15,14 @@ const user = {
 
 describe('Signup a user', () => {
 
-
-
-    beforeEach(async function() {
-        // runs before each test in this block
+    before(async function() {
         await records.deleteAllUsers();
     });
 
-    afterEach(async function() {
-        // runs before each test in this block
+    after(async function() {
         await records.deleteAllUsers();
     });
 
-    /**
-     * TEST USER SIGNUP
-     */
     it('Should add user to the db', (done) => {
         chai.request(server)
             .post('/api/v1/users/signup/')
@@ -37,24 +31,18 @@ describe('Signup a user', () => {
                 res.should.have.status(201);
                 done();
             });
-
     });
 });
-describe('Fetch all users', () => {
-    /**
-     * TEST GET ALL USERS
-     */
 
+
+describe('Fetch all users', () => {
     before(async function() {
-        // runs before each test in this block
         await records.deleteAllUsers();
     });
 
     after(async function() {
-        // runs before each test in this block
         await records.deleteAllUsers();
     });
-
 
     it('Should Fecth all the users', (done) => {
         chai.request(server)
@@ -64,32 +52,19 @@ describe('Fetch all users', () => {
                 done();
             });
     });
-
-
-
 });
 
 describe('Test fetch a single user', () => {
 
-
-
-
     before(async function() {
-        // runs before each test in this block
         await records.deleteAllUsers();
     });
 
     after(async function() {
-        // runs before each test in this block
         await records.deleteAllUsers();
     });
 
-
-
     it('Should Fecth a single user', (done) => {
-        /**
-         * TEST GET A SINGLE  USER
-         */
         chai.request(server)
             .post('/api/v1/users/signup/')
             .send(user)
@@ -113,7 +88,6 @@ describe('Test fetch a single user', () => {
 
 describe('Test update user', () => {
 
-
     const __user = {
         "firstName": "ui",
         "secondName": "fy",
@@ -126,19 +100,14 @@ describe('Test update user', () => {
 
 
     before(async function() {
-        // runs before each test in this block
         await records.deleteAllUsers();
     });
 
     after(async function() {
-        // runs before each test in this block
         await records.deleteAllUsers();
     });
 
     it('Should update a user', (done) => {
-        /**
-         * TEST GET A SINGLE  USER
-         */
         chai.request(server)
             .post('/api/v1/users/signup/')
             .send(user)
@@ -156,35 +125,23 @@ describe('Test update user', () => {
                             result.should.have.status(204);
                             done();
                         });
-
                 }
             });
     });
-
-
-
 });
-
-
 
 
 describe('Test delete user', () => {
 
-
     before(async function() {
-        // runs before each test in this block
         await records.deleteAllUsers();
     });
 
     after(async function() {
-        // runs before each test in this block
         await records.deleteAllUsers();
     });
 
     it('Should delete a user', (done) => {
-        /**
-         * TEST GET A SINGLE  USER
-         */
         chai.request(server)
             .post('/api/v1/users/signup/')
             .send(user)
@@ -202,23 +159,19 @@ describe('Test delete user', () => {
                 }
             });
     });
-
-
-
 });
 
 
-
 describe('Test signup with a string first name', () => {
+
     before(async function() {
-        // runs before each test in this block
         await records.deleteAllUsers();
     });
 
     after(async function() {
-        // runs before each test in this block
         await records.deleteAllUsers();
     });
+
     it('Should respond with a 400 status code while creating a user whose firstname is a number', (done) => {
         const _user = {
             "firstName": '67',
@@ -236,17 +189,17 @@ describe('Test signup with a string first name', () => {
                 res.should.have.status(400);
                 done();
             });
-
     });
 });
+
+
 describe('Creating a user with a short password', () => {
+
     before(async function() {
-        // runs before each test in this block
         await records.deleteAllUsers();
     });
 
     after(async function() {
-        // runs before each test in this block
         await records.deleteAllUsers();
     });
 
@@ -267,21 +220,19 @@ describe('Creating a user with a short password', () => {
                 res.should.have.status(400);
                 done();
             });
-
     });
 });
 
+
 describe('Test user signup with a string phone number', () => {
+
     before(async function() {
-        // runs before each test in this block
         await records.deleteAllUsers();
     });
 
     after(async function() {
-        // runs before each test in this block
         await records.deleteAllUsers();
     });
-
 
     it('Should respond with a 400 status code while creating a user whose phone number is a string', (done) => {
         const _user = {
@@ -304,16 +255,16 @@ describe('Test user signup with a string phone number', () => {
     });
 });
 
+
 describe('Test signup with non matching passwords', () => {
     before(async function() {
-        // runs before each test in this block
         await records.deleteAllUsers();
     });
 
     after(async function() {
-        // runs before each test in this block
         await records.deleteAllUsers();
     });
+
     it('Should respond with a 400 status code while creating a user whose passwords are not equal', (done) => {
         const _user = {
             "firstName": "hezron",
@@ -336,19 +287,13 @@ describe('Test signup with non matching passwords', () => {
 });
 
 
-
-
 describe('Test user signup with an invalid email', () => {
     before(async function() {
-        // runs before each test in this block
         await records.deleteAllUsers();
     });
-
     after(async function() {
-        // runs before each test in this block
         await records.deleteAllUsers();
     });
-
 
     it('Should respond with a 400 status code while creating a user whose email is invalid', (done) => {
         const _user = {
@@ -367,10 +312,5 @@ describe('Test user signup with an invalid email', () => {
                 res.should.have.status(400);
                 done();
             });
-
     });
-
-
-
-
 });
